Avalon

+++ Avalon/Avalon.Infra/Compare.cs
@@ -2,7 +2,7 @@ namespace Avalon.Infra;
 
 public class Compare : Any
 {
-    public virtual int Execute(object left, object right)
+    public virtual long Execute(object left, object right)
     {
         return 0;
     }
